# backend/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Simple Cheeseria API
  description: Displays cheese for a simple cheeseria POC app
  
# API paths
paths: 

  # All Cheeses
  /cheeses:
    get:
      description: Returns a list of cheeses
      responses:
        "200":
          description: Successfully returned a list of cheese
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cheese'
        "500":
          description: Server error

    post:
      description: Add a new cheese to the list of cheeses
      requestBody:
        description: The cheese data to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cheese'
      responses:
        "201":
          description: Successfully added new cheese
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cheese'
        "400":
          description: Invalid input
        "500":
          description: Server error

  # Individual cheese
  /cheeses/{id}:
    parameters:
      - in: path
        name: id
        required: true
        description: ID of the specific cheese
        schema:
          type: integer
          minimum: 1

    get:
      description: Returns a single cheese
      responses:
        "200":
          description: Successfully found and returned a cheese
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cheese'
        "400":
          description: Invalid cheese ID specified
        "404":
          description: No cheese with the specified ID was found
        "500":
          description: Server error
        default:
          description: Unexpected error

    put:
      description: Update specific cheese's data
      requestBody:
        description: The cheese data to overwrite with
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cheese'
      responses:
        "200":
          description: Successfully updated cheese data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cheese'
        "400":
          description: Invalid cheese ID specified
        "404":
          description: No cheese with the specified ID was found
        "500":
          description: Server error
        default:
          description: Unexpected error

    delete:
      description: Delete a cheese in the list of cheeses
      responses:
        "200":
          description: Successfully delete cheese data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cheese'
        "400":
          description: Invalid cheese ID specified
        "404":
          description: No cheese with the specified ID was found
        "500":
          description: Server error
        default:
          description: Unexpected error
 
# Components
components:

  # Generic definition for the cheese data
  schemas:
    Cheese:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        origin:
          type: string
        price:
          type: integer
        colour:
          type: string
        texture:
          type: string
        milk:
          type: string
        photo:
          type: string